Reorganize loader classes
